Changed totals to 32bit integers

+++ sw/airborne/modules/computer_vision/ground_detection_testing.c
@@ -75,9 +75,9 @@ int check_for_green(struct image_t *img, int right_corner_row, int right_corner_
     //pointer to buffer where image is stored
     uint8_t *buffer = img->buf;
 
-    double tot_lum;
-    double tot_cb;
-    double tot_cr;
+    uint32_t tot_lum;
+    uint32_t tot_cb;
+    uint32_t tot_cr;
 
     //Go through the pixels in the rectangle
     for (uint16_t y = right_corner_row;y < right_corner_row + LENGTH_SQUARE; y++) {
@@ -112,6 +112,10 @@ int check_for_green(struct image_t *img, int right_corner_row, int right_corner_
     double avg_lum = tot_lum/(LENGTH_SQUARE*WIDTH_RECT);
     double avg_cb = tot_cb/(LENGTH_SQUARE*WIDTH_RECT);
     double avg_cr = tot_cr/(LENGTH_SQUARE*WIDTH_RECT);
+
+    //printf("The average Y value is %lf \n", avg_lum);
+    //printf("The average U value is %lf \n", avg_cb);
+    //printf("The average V value is %lf \n", avg_cr);
 
     //If this average is outside the bounds make the pixel lighter and return 0
     if (!((avg_lum >= lum_min) && (avg_lum <= lum_max) &&
